Refactor dialogs and popovers into their own file so layout can stop worrying about it.
